Improve eror handling
